refactor: Split more code, updated documentation, added TODOs to show what should me moved to core.
